requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
playwright==1.51.0
beautifulsoup4==4.13.3
requests==2.32.3
python-dotenv==1.1.0
fake-useragent==1.4.0
lxml==5.1.0
openai==1.12.0
price-parser==0.3.4
aiohttp==3.9.3
python-multipart==0.0.9
pydantic==2.6.3
httpx==0.26.0
selenium==4.18.1
webdriver-manager==4.0.1 